+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/VenusContext.java
@@ -51,6 +51,9 @@ public class VenusContext {
     @Getter @Setter
     private boolean addVerifyPlugin;
 
+    @Getter @Setter
+    private boolean addJunit5;
+
     public void setExcludeXmlApis( boolean excludeXmlApis ) {
         if ( excludeXmlApis ) {
             this.setAddExclusions( "xml-apis:xml-apis" );
@@ -63,6 +66,10 @@ public class VenusContext {
         this.extensions = extensions;
         this.errors = new ArrayList<>();
         this.modules = new HashSet<>();
+        this.addDocFacace = true;
+        this.force = false;
+        this.addVerifyPlugin = true;
+        this.addJunit5 = false;
     }
 
     public String getGroupId() {